+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 
+ * Add builds for Android to zlib sample project ([issue #5](https://github.com/bytedeco/gradle-javacpp/issues/5))
+
 ### September 9, 2020 version 1.5.4
  * Initial release